LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/inline-size-containment-vertical-rl-expected.txt


FAIL inline-size containment only assert_equals: expected 400 but got 20
